Add and start general_programming.py
general_programming.py will house answers for the General Programming section
of questions.

Created dict_to_list function to answer question 1. Given the example there is
a way to make the function faster, but the function would not be applicable to
dictionaries with different numbers of keys or length of values.